requirements.txt
alembic==1.0.0
asn1crypto==0.24.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.3.1
dominate==2.3.4
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==3.2.4
Flask-JWT==0.3.2
Flask-JWT-Extended==3.13.1
Flask-Login==0.4.1
flask-marshmallow==0.9.0
Flask-Migrate==2.2.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
marshmallow==2.15.6
marshmallow-sqlalchemy==0.14.1
phonenumbers==8.9.14
pkg-resources==0.0.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pycparser==2.19
PyJWT==1.6.4
PySocks==1.6.8
python-dateutil==2.7.3
python-dotenv==0.9.1
python-editor==1.0.3
pytz==2018.5
requests==2.19.1
six==1.11.0
SQLAlchemy==1.2.12
SQLAlchemy-Utils==0.33.5
twilio==6.18.0
urllib3==1.23
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
